Resolves broken tests on local and adds GH actions pipeline (#63)
* Resolves broken tests on local and adds gh actions pipeline

* additional cleanup

* remove sniffer config

* remove static analyzers for now

* updates pipeline to deal with cake4 requiring php 7.4

* cleanup and emit deprecation warning on deprecated method

* adds elastic search test into gh action pipeline with a custom integration test, uses mocks for testing elastic locally, some other cleanup

* tests pipeline without elastic to debug

* resolve timezone bug

* resolve timezone bug

* resolve timezone bug

* resolve timezone bug

* resolve timezone bug

* resolve timezone bug

* resolve timezone bug

* resolve timezone bug

* resolve timezone bug

* skip timezone bug and enable elastic again

* debugging this: https://github.com/elastic/elastic-github-actions/issues/23

* bad copy pasta

* debug datetime issue

* oh well

* oh well 2

* adds #[\ReturnTypeWillChange] to suppress deprecation warnings 8.0/8.1

* cleanup

* docblock updates

* re-adds integration test and removes a deprecation warning in php 8.1

* testing pipeline

* fix test namespaces

* try getting elastc running in pipeline again using setup from cakephp/elastic

* try getting elastc running in pipeline again using setup from cakephp/elastic

* try getting elastc running in pipeline again using setup from cakephp/elastic

* fix names
