+++ composer.json
@@ -4,8 +4,9 @@
     "type": "cakephp-plugin",
     "license": "MIT",
     "require": {
-        "php": ">=7.2",
-        "cakephp/orm": "^4.0"
+        "php": "^7.2|^8.0",
+        "cakephp/orm": "^4.0",
+        "ext-json": "*"
     },
     "require-dev": {
         "phpunit/phpunit": "^8.0",
@@ -25,8 +26,11 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "AuditStash\\Test\\": "tests",
+            "AuditStash\\Test\\": "tests/",
             "Cake\\Test\\": "./vendor/cakephp/cakephp/tests"
         }
+    },
+    "scripts": {
+        "test": "phpunit --colors=always"
     }
 }
