+++ src/EventFactory.php
@@ -18,7 +18,8 @@ class EventFactory
      * converts it into an AuditStash\EventInterface object.
      *
      * @param array $data The array data from elastic search
-     * @return AuditStash\EventInterface
+     * @return \AuditStash\EventInterface
+     * @throws \ReflectionException
      */
     public function create(array $data)
     {
