+++ src/Meta/ApplicationMetadata.php
@@ -40,10 +40,9 @@ class ApplicationMetadata implements EventListenerInterface
     }
 
     /**
-     * Enriches all of the passed audit logs to add the request
-     * info metadata.
+     * Enriches all the passed audit logs to add the request info metadata.
      *
-     * @param Event The AuditStash.beforeLog event
+     * @param Event $event The AuditStash.beforeLog event
      * @param array $logs The audit log event objects
      * @return void
      */
