+++ src/Meta/RequestMetadata.php
@@ -49,10 +49,9 @@ class RequestMetadata implements EventListenerInterface
     }
 
     /**
-     * Enriches all of the passed audit logs to add the request
-     * info metadata.
+     * Enriches all the passed audit logs to add the request info metadata.
      *
-     * @param Event The AuditStash.beforeLog event
+     * @param Event $event The AuditStash.beforeLog event
      * @param array $logs The audit log event objects
      * @return void
      */
