+++ src/Persister/ElasticSearchPersister.php
@@ -6,17 +6,18 @@ use AuditStash\Exception;
 use AuditStash\PersisterInterface;
 use Cake\Datasource\ConnectionManager;
 use Cake\ElasticSearch\Datasource\Connection;
+use Elastica\Client;
 use Elastica\Document;
 
 /**
- * Implementes audit logs events persisting using Elasticsearch.
+ * Implements audit logs events persisting using Elasticsearch.
  */
 class ElasticSearchPersister implements PersisterInterface
 {
     /**
      * The client or connection to Elasticsearch.
      *
-     * @var Cake\ElasticSearch\Datasource\Connection
+     * @var \Cake\ElasticSearch\Datasource\Connection
      */
     protected $connection;
 
@@ -49,7 +50,9 @@ class ElasticSearchPersister implements PersisterInterface
      * - index: The Elasticsearch index to store documents
      * - type: The Elasticsearch mapping type of documents
      *
+     * @param array $options
      * @return void
+     * @throws Exception
      */
     public function __construct($options = [])
     {
@@ -73,23 +76,24 @@ class ElasticSearchPersister implements PersisterInterface
     }
 
     /**
-     * Persists all of the audit log event objects that are provided.
+     * Persists all the audit log event objects that are provided.
      *
-     * @param array $auditLogs An array of EventInterface objects
+     * @param \AuditStash\EventInterface[] $auditLogs An array of EventInterface objects
      * @return void
      */
     public function logEvents(array $auditLogs)
     {
-        $client = $this->getConnection();
         $documents = $this->transformToDocuments($auditLogs);
 
+        $connection = $this->getConnection();
+        $client = $connection->getDriver();
         $client->addDocuments($documents);
     }
 
     /**
      * Transforms the EventInterface objects to Elastica Documents.
      *
-     * @param array $auditLogs An array of EventInterface objects.
+     * @param \AuditStash\EventInterface[] $auditLogs An array of EventInterface objects.
      * @return array
      */
     protected function transformToDocuments($auditLogs)
@@ -124,7 +128,7 @@ class ElasticSearchPersister implements PersisterInterface
      * id in elastic search. Only enable this feature if you know that your transactions are
      * only comprised of a single event log per commit.
      *
-     * @param bool $use Whether or not to copy the transactionId as the document id
+     * @param bool $use Whether to copy the transactionId as the document id
      * @return void
      */
     public function reuseTransactionId($use = true)
@@ -135,7 +139,7 @@ class ElasticSearchPersister implements PersisterInterface
     /**
      * Sets the client connection to elastic search.
      *
-     * @param Elastica\Client $connection The conneciton to elastic search
+     * @param \Cake\ElasticSearch\Datasource\Connection $connection The connection to elastic search
      * @return $this
      */
     public function setConnection(Connection $connection)
@@ -150,7 +154,7 @@ class ElasticSearchPersister implements PersisterInterface
      *
      * If connection is not defined, create a new one.
      *
-     * @return Elastica\Client
+     * @return \Cake\ElasticSearch\Datasource\Connection
      */
     public function getConnection()
     {
@@ -165,12 +169,13 @@ class ElasticSearchPersister implements PersisterInterface
      * Sets the client connection to elastic search when passed.
      * If no arguments are provided, it returns the current connection.
      *
+     * @param \Elastica\Client|null $connection The connection to elastic search
+     * @return \Elastica\Client
      * @deprecated Use getConnection()/setConnection() instead
-     * @param Elastica\Client $connection The conneciton to elastic search
-     * @return Elastica\Client
      */
     public function connection(Client $connection = null)
     {
+        deprecationWarning('Use getConnection()/setConnection() instead');
         if ($connection !== null) {
             return $this->setConnection($connection);
         }
