+++ src/Persister/TablePersister.php
@@ -171,7 +171,7 @@ class TablePersister implements PersisterInterface
      */
     public function logEvents(array $auditLogs)
     {
-        $PersisterTable = $this->getTable();
+        $persisterTable = $this->getTable();
 
         $serializeFields = $this->getConfig('serializeFields');
         $primaryKeyExtractionStrategy = $this->getConfig('primaryKeyExtractionStrategy');
@@ -186,9 +186,9 @@ class TablePersister implements PersisterInterface
                 $log, $extractMetaFields, $unsetExtractedMetaFields, $serializeFields
             );
 
-            $persisterEntity = $PersisterTable->newEntity($fields);
+            $persisterEntity = $persisterTable->newEntity($fields);
 
-            if (!$PersisterTable->save($persisterEntity) &&
+            if (!$persisterTable->save($persisterEntity) &&
                 $logErrors
             ) {
                 $this->log($this->toErrorLog($persisterEntity));
